load_balanced_ec2_webserver/requirements.txt
aws-cdk.core
aws-cdk.aws_s3
aws-cdk.aws_ec2
aws-cdk.aws_iam
aws-cdk.aws_autoscaling
aws-cdk.aws_elasticloadbalancingv2
